GH-252: masq seems to be encoding the command properly
