+++ masq/src/commands/mod.rs
@@ -1,6 +1,6 @@
 // Copyright (c) 2019-2020, MASQ (https://masq.ai) and/or its affiliates. All rights reserved.
 
-pub mod commands;
+pub mod commands_common;
 pub mod setup_command;
-pub mod start_command;
 pub mod shutdown_command;
+pub mod start_command;
