+++ masq/src/schema.rs
@@ -1,10 +1,10 @@
 // Copyright (c) 2019-2020, MASQ (https://masq.ai) and/or its affiliates. All rights reserved.
+use crate::commands::setup_command::setup_subcommand;
+use crate::commands::shutdown_command::shutdown_subcommand;
+use crate::commands::start_command::start_subcommand;
 use clap::{crate_description, crate_version, App, AppSettings, Arg};
 use lazy_static::lazy_static;
 use masq_lib::constants::{DEFAULT_UI_PORT, HIGHEST_USABLE_PORT, LOWEST_USABLE_INSECURE_PORT};
-use crate::commands::setup_command::setup_subcommand;
-use crate::commands::start_command::start_subcommand;
-use crate::commands::shutdown_command::shutdown_subcommand;
 
 lazy_static! {
     static ref UI_PORT_HELP: String = format!(
