+++ masq_lib/src/utils.rs
@@ -38,3 +38,129 @@ pub fn find_free_port() -> u16 {
 pub fn localhost() -> IpAddr {
     IpAddr::V4(Ipv4Addr::LOCALHOST)
 }
+
+#[allow(clippy::needless_range_loop)]
+pub fn index_of<T>(haystack: &[T], needles: &[T]) -> Option<usize>
+where
+    T: PartialEq,
+{
+    if needles.is_empty() {
+        return None;
+    }
+    for h in 0..haystack.len() {
+        let mut mismatch = false;
+        for n in 0..needles.len() {
+            let i = h + n;
+            if i >= haystack.len() {
+                mismatch = true;
+                break;
+            }
+            if haystack[i] != needles[n] {
+                mismatch = true;
+                break;
+            }
+        }
+        if !mismatch {
+            return Some(h);
+        }
+    }
+    None
+}
+
+pub fn index_of_from<T>(haystack: &[T], needle: &T, start_at: usize) -> Option<usize>
+where
+    T: PartialEq,
+{
+    let mut index = start_at;
+    while index < haystack.len() && (haystack[index] != *needle) {
+        index += 1;
+    }
+    if index >= haystack.len() {
+        None
+    } else {
+        Some(index)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn index_of_fails_to_find_nonexistent_needle_in_haystack() {
+        let result = index_of("haystack".as_bytes(), "needle".as_bytes());
+
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn index_of_finds_needle_at_beginning_of_haystack() {
+        let result = index_of("haystack haystack".as_bytes(), "haystack".as_bytes());
+
+        assert_eq!(result, Some(0));
+    }
+
+    #[test]
+    fn index_of_finds_needle_at_end_of_haystack() {
+        let result = index_of("needle haystack".as_bytes(), "haystack".as_bytes());
+
+        assert_eq!(result, Some(7));
+    }
+
+    #[test]
+    fn index_of_fails_to_find_nonempty_needle_in_empty_haystack() {
+        let result = index_of("".as_bytes(), "needle".as_bytes());
+
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn index_of_returns_none_for_empty_needle() {
+        let result = index_of("haystack".as_bytes(), "".as_bytes());
+
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn index_of_fails_to_find_needle_that_ends_past_end_of_haystack() {
+        let result = index_of("haystack needl".as_bytes(), "needle".as_bytes());
+
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn index_of_from_fails_to_find_nonexistent_needle_in_haystack() {
+        let haystack = vec![true, true, true, true];
+
+        let result = index_of_from(&haystack, &false, 0);
+
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn index_of_from_fails_to_find_needle_in_empty_haystack() {
+        let haystack: Vec<i32> = vec![];
+
+        let result = index_of_from(&haystack, &-42, 0);
+
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn index_of_from_finds_needle_at_beginning_of_search() {
+        let haystack = vec![8, 7, 8, 3];
+
+        let result = index_of_from(&haystack, &8, 2);
+
+        assert_eq!(result, Some(2));
+    }
+
+    #[test]
+    fn index_of_from_finds_needle_at_end_of_haystack() {
+        let haystack = vec![8, 7, 8, 3];
+
+        let result = index_of_from(&haystack, &3, 0);
+
+        assert_eq!(result, Some(3));
+    }
+}
