+++ node/src/proxy_server/http_protocol_pack.rs
@@ -3,9 +3,9 @@ use crate::proxy_server::protocol_pack::{Host, ProtocolPack, ServerImpersonator}
 use crate::proxy_server::server_impersonator_http::ServerImpersonatorHttp;
 use crate::sub_lib::cryptde::PlainData;
 use crate::sub_lib::proxy_server::ProxyProtocol;
-use crate::sub_lib::utils::index_of;
 use lazy_static::lazy_static;
 use masq_lib::constants::HTTP_PORT;
+use masq_lib::utils::index_of;
 use regex::Regex;
 
 lazy_static! {
