+++ node/src/sub_lib/http_packet_framer.rs
@@ -3,9 +3,9 @@ use crate::sub_lib::framer::FramedChunk;
 use crate::sub_lib::framer::Framer;
 use crate::sub_lib::framer_utils;
 use crate::sub_lib::logger::Logger;
-use crate::sub_lib::utils::index_of;
-use crate::sub_lib::utils::index_of_from;
 use crate::sub_lib::utils::to_string;
+use masq_lib::utils::index_of;
+use masq_lib::utils::index_of_from;
 use regex::Regex;
 use std::fmt;
 use std::fmt::Debug;
